Review notes applied.
