+++ examples/dbclient/pokemons/src/main/java/io/helidon/examples/dbclient/pokemons/PokemonMapper.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.helidon.examples.dbclient.pokemons.mapper;
+package io.helidon.examples.dbclient.pokemons;
 
 import java.util.ArrayList;
 import java.util.HashMap;
