+++ examples/dbclient/pokemons/src/main/java/io/helidon/examples/dbclient/pokemons/PokemonMapperProvider.java
@@ -13,8 +13,28 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+package io.helidon.examples.dbclient.pokemons;
+
+import java.util.Optional;
+
+import javax.annotation.Priority;
+
+import io.helidon.dbclient.DbMapper;
+import io.helidon.dbclient.spi.DbMapperProvider;
 
 /**
- * Database statements to POJO mapping.
+ * Provides pokemon mappers.
  */
-package io.helidon.examples.dbclient.pokemons.mapper;
+@Priority(1000)
+public class PokemonMapperProvider implements DbMapperProvider {
+    private static final PokemonMapper MAPPER = new PokemonMapper();
+
+    @SuppressWarnings("unchecked")
+    @Override
+    public <T> Optional<DbMapper<T>> mapper(Class<T> type) {
+        if (type.equals(Pokemon.class)) {
+            return Optional.of((DbMapper<T>) MAPPER);
+        }
+        return Optional.empty();
+    }
+}
