+++ examples/dbclient/pokemons/src/main/resources/META-INF/services/io.helidon.dbclient.spi.DbMapperProvider
@@ -14,4 +14,4 @@
 # limitations under the License.
 #
 
-io.helidon.examples.dbclient.pokemons.mapper.PokemonMapperProvider
+io.helidon.examples.dbclient.pokemons.PokemonMapperProvider
